feat(user): implement user creation logic with validation and email verification
